mini-rpa/agent/requirements.txt
flask
requests
python-dotenv
cryptography
uvicorn
keyring
pydantic
fastapi